Read an enum's discriminants as rustc computes them
`VariantDiscr::Relative(i)` is the distance from the nearest preceding explicit
discriminant, not a discriminant, so every variant following an explicit one was
encoded with a value that the `SwitchInt` of a `match` never compares against:
its arm ran under an unsatisfiable path assumption and verified vacuously,
accepting programs that panic. Reading an explicit discriminant went through
`ScalarInt::to_u32`, which asserts a four-byte tag, so an enum carrying one bugs
out on every tag type but `i32`/`u32` — `isize` by default, and `i8` for
`core::cmp::Ordering`, which any `#[derive(PartialOrd)]` in the crate pulls in.

Take the absolute discriminants rustc already computes, sign-extended to the
width their tag type declares so they agree with the `SwitchInt` targets, and
carry them as `i64` from there on.

Fixes #126

+++ src/refine/env.rs
@@ -770,7 +770,7 @@ where
                 field_ty.instantiate_ty_params(ty.args.clone());
                 let guarded_field_ty = field_ty.guarded(
                     chc::Term::var(discr_var.into())
-                        .equal_to(chc::Term::int(variant_def.discr as i64))
+                        .equal_to(chc::Term::int(variant_def.discr))
                         .into(),
                 );
                 self.bind_impl(x.into(), guarded_field_ty.boxed(), depth);
